+++ packages/client-reconciliation/RELEASES.md
@@ -89,3 +89,13 @@ Based on:
 - [typescript v1.0.0-alpha.6] packages/client-reconciliation
 ### Releases
 - [NPM v1.0.0-alpha.6] https://www.npmjs.com/package/@formance/sdk-reconciliation/v/1.0.0-alpha.6 - packages/client-reconciliation
+
+## 2024-07-30 00:16:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.348.1 (2.380.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.7] packages/client-reconciliation
+### Releases
+- [NPM v1.0.0-alpha.7] https://www.npmjs.com/package/@formance/sdk-reconciliation/v/1.0.0-alpha.7 - packages/client-reconciliation